fix(editor): stabilize quote list editing
Refs #224
